refactor: move pusher types to workflow/variables/types.py
Move ModelArtifact, AssembledModel, PusherResult from pusher/types.py
to workflow/variables/types.py — the canonical location for variable
types shared across workflow tasks per the CanvasFlex design doc.

- workflow/variables/types.py — new canonical location
- workflow/variables/__init__.py — exports the three types; wraps
  DatasetVariable import in contextlib.suppress(ImportError) so the
  lightweight types remain importable without pyspark/pandas installed
- pusher/types.py — thin re-export shim pointing to new location
- tests/test_types.py — imports updated to workflow.variables.types
